added framework for async unit tests

+++ src/test/java/nl/cerios/reactive/chocolate/factory/model/ChocoNutTest.kt
@@ -16,7 +16,7 @@ internal class ChocoNutTest {
     val json = chocoNut1.toJson()
     val chocoNut2 = ChocoNut.fromJson(json)
 
-    assertEquals(quality, json.getFloat("quality"))
+    assertEquals(quality, json.getDouble("quality"))
     assertEquals(flavor.name, json.getString("flavor"))
     assertEquals("""{"quality":0.25,"flavor":"MILK"}""", json.encode())
     assertEquals(chocoNut1, chocoNut2)
